feat: 权限交接
