+++ frontend/src/views/transfer/system-super-manager.css
@@ -1,0 +1,90 @@
+.iam-transfer-system-wrapper {
+    /* min-height: 328px; */
+    margin-bottom: 20px;
+    .transfer-system-content {
+        background-color: #fff;
+        border-radius: 2px;
+        border: 1px solid #fff;
+        box-shadow: 0px 1px 2px 0px rgba(49, 50, 56, .1);
+        &:hover {
+            box-shadow: 0px 2px 4px 0px rgba(49, 50, 56, .15);
+        }
+        .checkbox-cell-wrapper {
+            .cell {
+                padding-left: 0;
+            }
+        }
+        .header {
+            display: flex;
+            justify-content: flex-start;
+            position: relative;
+            padding: 0 30px;
+            height: 40px;
+            line-height: 40px;
+            font-size: 14px;
+            color: #63656e;
+            cursor: pointer;
+            user-select: none;
+            .expanded-icon {
+                position: absolute;
+                top: 13px;
+                left: 10px;
+            }
+            .title {
+                font-weight: 500;
+                cursor: pointer;
+                color: #313238;
+            }
+            .sub-title {
+                margin-left: 17px;
+                .reason {
+                    font-size: 12px;
+                    color: #979ba5;
+                }
+            }
+        }
+        .content {
+            position: relative;
+            .expand-action {
+                padding: 0 0 5px 0;
+                margin-top: -10px;
+                width: 100%;
+                line-height: 16px;
+                font-size: 12px;
+                color: #979ba5;
+                text-align: center;
+                cursor: pointer;
+                i {
+                    font-size: 16px;
+                }
+            }
+            .slot-content {
+                padding: 0 30px 20px 30px;
+                min-height: 60px;
+            }
+        }
+        .member-item {
+            position: relative;
+            display: inline-block;
+            margin: 0 6px 6px 0;
+            padding: 0 10px;
+            line-height: 22px;
+            background: #f5f6fa;
+            border: 1px solid #dcdee5;
+            border-radius: 2px;
+            font-size: 12px;
+            .member-name {
+                display: inline-block;
+                max-width: 200px;
+                line-height: 17px;
+                overflow: hidden;
+                text-overflow: ellipsis;
+                white-space: nowrap;
+                vertical-align: text-top;
+                .count {
+                    color: #c4c6cc;
+                }
+            }
+        }
+    }
+}
